move MatchedFieldList to search/matches directory
